+++ examples/03-dev/demo/project/src/test/index/js/wel.jsx
@@ -1,9 +1,11 @@
 
+import { Button } from 'antd';
+
 var React = require('react');
 
 class Welcome extends React.Component {
     render() {
-        return <h1>Hello, {this.props.name}</h1>;
+        return <Button>Hello, {this.props.name}</Button>;
     }
 }