+++ examples/03-dev/index.js
@@ -1,5 +1,5 @@
 
 require('../../util/change_cwd_to')(__dirname + '/demo');
 
-// require('../../util/exec')('lila dev test-5/index -l senn -o');
-require('../../util/exec')('lila dev test/index');
+require('../../util/exec')('lila dev test/index -l senn -o');
+// require('../../util/exec')('lila dev test/index');